Added query about meaning of 'threadsafe' (from ffi mailing list 29 Nov 2002).
It is not clear from the report what 'threadsafe' means but my best
guess suggests that the report is wrong to say that implementations
which support a single Haskell thread may safely ignore the
'threadsafe' keyword.  Indeed, it seems that such implementations
should either add extra locking or reject 'threadsafe' ffi
declarations.